Move softmax.py frm pytorch/ to pytorch/d_p_a
