+++ transformer_engine/pytorch/dot_product_attention/softmax.py
